requirements.txt
numpy
pyserial
fastapi
uvicorn[standard]
requests
